liran review
