+++ mlrun/utils/helpers.py
@@ -46,7 +46,6 @@ import semver
 import yaml
 from dateutil import parser
 from pandas import Timedelta, Timestamp
-from timelength import TimeLength
 from yaml.representer import RepresenterError
 
 import mlrun
@@ -2312,20 +2311,3 @@ def encode_user_code(
             "Consider using `with_source_archive` to add user code as a remote source to the function."
         )
     return encoded
-
-
-def time_string_to_seconds(time_str: str, min_seconds: int = 60) -> Optional[int]:
-    if not time_str:
-        return None
-
-    if time_str == "-1":
-        return -1
-
-    parsed_length = TimeLength(time_str, strict=True)
-    total_seconds = parsed_length.to_seconds()
-    if total_seconds < min_seconds:
-        raise ValueError(
-            f"Invalid time string {time_str}, must be at least {min_seconds=}"
-        )
-
-    return total_seconds
