+++ requirements.txt
@@ -42,6 +42,5 @@ orjson>=3.9.15, <4
 mlrun-pipelines-kfp-common~=0.5.5
 mlrun-pipelines-kfp-v1-8~=0.5.4
 docstring_parser~=0.16
-timelength~=1.1
 # TODO: Remove it after moving notificaions to the server side(ML-8069)
 aiosmtplib~=3.0
