+++ server/py/services/api/launcher.py
@@ -18,6 +18,7 @@ from dependency_injector import containers, providers
 
 import mlrun.common.constants as mlrun_constants
 import mlrun.common.db.sql_session
+import mlrun.common.runtimes.constants
 import mlrun.common.schemas.schedule
 import mlrun.config
 import mlrun.execution
@@ -414,6 +415,7 @@ class ServerSideLauncher(launcher.BaseLauncher):
             )
 
         self._validate_state_thresholds(run.spec.state_thresholds)
+        self._validate_retry(runtime.kind, run.spec.retry)
 
         if (
             mlrun.runtimes.RuntimeKinds.requires_image_name_for_execution(runtime.kind)
@@ -455,11 +457,43 @@ class ServerSideLauncher(launcher.BaseLauncher):
                 )
 
             try:
-                mlrun.utils.helpers.time_string_to_seconds(threshold)
+                framework.utils.helpers.time_string_to_seconds(threshold)
             except Exception as exc:
                 raise mlrun.errors.MLRunInvalidArgumentError(
                     f"Threshold '{threshold}' for state '{state}' is not a valid timelength string. "
                     f"Error: {mlrun.errors.err_to_str(exc)}"
+                ) from exc
+
+    @staticmethod
+    def _validate_retry(runtime_kind: str, retry: Optional["mlrun.model.Retry"]):
+        if retry is None or not retry.count:
+            return
+
+        if retry.count is not None and retry.count < 0:
+            raise mlrun.errors.MLRunInvalidArgumentError(
+                f"Retry count must be at least 0, got {retry.count}"
+            )
+
+        if runtime_kind not in mlrun.runtimes.RuntimeKinds.retriable_runtimes():
+            raise mlrun.errors.MLRunInvalidArgumentError(
+                f"Retry is not supported for runtime kind {runtime_kind}, "
+                "supported kinds are: "
+                f"{mlrun.runtimes.RuntimeKinds.retriable_runtimes()}"
+            )
+
+        backoff = retry.backoff
+        if backoff is not None and backoff.base_delay is not None:
+            min_base_delay = mlrun.mlconf.function.spec.retry.backoff.min_base_delay
+            min_base_delay_seconds = framework.utils.helpers.time_string_to_seconds(
+                min_base_delay, 0
+            )
+            try:
+                framework.utils.helpers.time_string_to_seconds(
+                    backoff.base_delay, min_base_delay_seconds
+                )
+            except ValueError as exc:
+                raise mlrun.errors.MLRunInvalidArgumentError(
+                    f"Retry backoff base_delay must be at least {min_base_delay}, got {backoff.base_delay}"
                 ) from exc
