+++ server/py/services/api/tests/unit/runtime_handlers/test_kubejob.py
@@ -592,7 +592,7 @@ class TestKubejobRuntimeHandler(TestRuntimeHandlerBase):
         pending_scheduled_pod.status.start_time = datetime.now(
             timezone.utc
         ) - timedelta(
-            seconds=mlrun.utils.helpers.time_string_to_seconds(
+            seconds=framework.utils.helpers.time_string_to_seconds(
                 mlrun.mlconf.function.spec.state_thresholds.default.pending_scheduled
             )
         )
@@ -630,7 +630,7 @@ class TestKubejobRuntimeHandler(TestRuntimeHandlerBase):
             PodPhases.running,
         )
         running_overtime_pod.status.start_time = datetime.now(timezone.utc) - timedelta(
-            seconds=mlrun.utils.helpers.time_string_to_seconds(
+            seconds=framework.utils.helpers.time_string_to_seconds(
                 mlrun.mlconf.function.spec.state_thresholds.default.executing
             )
         )
@@ -666,7 +666,7 @@ class TestKubejobRuntimeHandler(TestRuntimeHandlerBase):
         image_pull_backoff_pod.status.start_time = datetime.now(
             timezone.utc
         ) - timedelta(
-            seconds=mlrun.utils.helpers.time_string_to_seconds(
+            seconds=framework.utils.helpers.time_string_to_seconds(
                 mlrun.mlconf.function.spec.state_thresholds.default.image_pull_backoff
             )
         )
