+++ server/py/services/api/tests/unit/runtime_handlers/test_mpijob.py
@@ -26,6 +26,7 @@ import mlrun.utils.helpers
 from mlrun.common.runtimes.constants import PodPhases, RunStates
 from mlrun.runtimes import RuntimeKinds
 
+import framework.utils.helpers
 from framework.utils.singletons.db import get_db
 from framework.utils.singletons.k8s import get_k8s_helper
 from services.api.runtime_handlers import get_runtime_handler
@@ -376,7 +377,7 @@ class TestMPIjobRuntimeHandler(TestRuntimeHandlerBase):
         """
         # set big debouncing interval to avoid having to mock resources for all the runs on every monitor cycle
         mlrun.mlconf.monitoring.runs.missing_runtime_resources_debouncing_interval = (
-            mlrun.utils.helpers.time_string_to_seconds(
+            framework.utils.helpers.time_string_to_seconds(
                 mlrun.mlconf.function.spec.state_thresholds.default.executing
             )
             * 2
@@ -392,7 +393,7 @@ class TestMPIjobRuntimeHandler(TestRuntimeHandlerBase):
                 image_pull_backoff_job_uid,
                 datetime.now(timezone.utc)
                 - timedelta(
-                    seconds=mlrun.utils.helpers.time_string_to_seconds(
+                    seconds=framework.utils.helpers.time_string_to_seconds(
                         mlrun.mlconf.function.spec.state_thresholds.default.image_pull_backoff
                     )
                 ),
@@ -401,7 +402,7 @@ class TestMPIjobRuntimeHandler(TestRuntimeHandlerBase):
                 running_long_uid,
                 datetime.now(timezone.utc)
                 - timedelta(
-                    seconds=mlrun.utils.helpers.time_string_to_seconds(
+                    seconds=framework.utils.helpers.time_string_to_seconds(
                         mlrun.mlconf.function.spec.state_thresholds.default.executing
                     )
                 ),
@@ -553,7 +554,7 @@ class TestMPIjobRuntimeHandler(TestRuntimeHandlerBase):
     def test_state_thresholds_pending_states(self, db: Session, client: TestClient):
         # set big debouncing interval to avoid having to mock resources for all the runs on every monitor cycle
         mlrun.mlconf.monitoring.runs.missing_runtime_resources_debouncing_interval = (
-            mlrun.utils.helpers.time_string_to_seconds(
+            framework.utils.helpers.time_string_to_seconds(
                 mlrun.mlconf.function.spec.state_thresholds.default.pending_scheduled
             )
             * 2
@@ -570,7 +571,7 @@ class TestMPIjobRuntimeHandler(TestRuntimeHandlerBase):
                 pending_scheduled_stale_uid,
                 datetime.now(timezone.utc)
                 - timedelta(
-                    seconds=mlrun.utils.helpers.time_string_to_seconds(
+                    seconds=framework.utils.helpers.time_string_to_seconds(
                         mlrun.mlconf.function.spec.state_thresholds.default.pending_scheduled
                     )
                 ),
