+++ server/py/services/api/tests/unit/runtime_handlers/test_sparkjob.py
@@ -27,6 +27,7 @@ import mlrun.utils.helpers
 from mlrun.common.runtimes.constants import PodPhases, RunStates
 from mlrun.runtimes import RuntimeKinds
 
+import framework.utils.helpers
 from framework.utils.singletons.db import get_db
 from framework.utils.singletons.k8s import get_k8s_helper
 from services.api.runtime_handlers import get_runtime_handler
@@ -401,7 +402,7 @@ class TestSparkjobRuntimeHandler(TestRuntimeHandlerBase):
         stale_run_name = "my-spark-stale"
         new_run_name = "my-spark-new"
 
-        threshold_in_seconds = mlrun.utils.helpers.time_string_to_seconds(
+        threshold_in_seconds = framework.utils.helpers.time_string_to_seconds(
             getattr(
                 mlrun.mlconf.function.spec.state_thresholds.default,
                 threshold_state,
