+++ server/py/services/api/tests/unit/test_launcher.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import pathlib
+import re
 import unittest.mock
 from contextlib import nullcontext as does_not_raise
 
@@ -29,6 +30,10 @@ from mlrun.config import Config
 import framework.utils.clients.iguazio
 import services.api.launcher
 import services.api.tests.unit.api.utils
+
+assets_path = pathlib.Path(__file__).parent / "assets"
+func_path = assets_path / "sample_function.py"
+handler = "hello_word"
 
 
 @pytest.mark.parametrize(
@@ -146,9 +151,6 @@ def test_new_function_args_with_default_image_pull_secret(
     db: sqlalchemy.orm.Session, client: TestClient
 ):
     project = "some-project"
-    assets_path = pathlib.Path(__file__).parent / "assets"
-    func_path = assets_path / "sample_function.py"
-    handler = "hello_word"
     services.api.tests.unit.api.utils.create_project(client, project)
 
     mlrun.mlconf.function.spec.image_pull_secret = Config(
@@ -189,3 +191,111 @@ def test_new_function_args_with_default_image_pull_secret(
         runtime.spec.image_pull_secret
         == mlrun.mlconf.function.spec.image_pull_secret.default
     )
+
+
+@pytest.mark.parametrize(
+    "count, base_delay, default_base_delay, min_base_delay, expectation",
+    [
+        (None, None, "30s", "30s", does_not_raise()),
+        (
+            1,
+            "29s",
+            "30s",
+            "30s",
+            pytest.raises(
+                mlrun.errors.MLRunInvalidArgumentError,
+                match="Retry backoff base_delay must be at least 30s, got 29s",
+            ),
+        ),
+        (
+            1,
+            "31s",
+            "30s",
+            "5m",
+            pytest.raises(
+                mlrun.errors.MLRunInvalidArgumentError,
+                match="Retry backoff base_delay must be at least 5m, got 31s",
+            ),
+        ),
+        (3, None, "30s", "30s", does_not_raise()),
+        (3, "1 min", "30s", "30s", does_not_raise()),
+        (
+            -1,
+            None,
+            "30s",
+            "30s",
+            pytest.raises(
+                mlrun.errors.MLRunInvalidArgumentError,
+                match="Retry count must be at least 0, got -1",
+            ),
+        ),
+    ],
+)
+def test_validate_run_retry(
+    count, base_delay, default_base_delay, min_base_delay, expectation
+):
+    mlrun.mlconf.function.spec.retry.backoff.default_base_delay = default_base_delay
+    mlrun.mlconf.function.spec.retry.backoff.min_base_delay = min_base_delay
+    launcher = services.api.launcher.ServerSideLauncher(
+        auth_info=mlrun.common.schemas.AuthInfo()
+    )
+    runtime = mlrun.code_to_function(
+        name="test", kind="job", filename=str(func_path), handler=handler
+    )
+
+    retry = None
+    if count or base_delay:
+        retry = {}
+        if count is not None:
+            retry["count"] = count
+
+        if base_delay is not None:
+            retry["backoff"] = {
+                "base_delay": base_delay,
+            }
+
+    run = mlrun.run.RunObject(
+        spec=mlrun.model.RunSpec(
+            retry=retry,
+        ),
+    )
+    assert run.spec.retry.count == (count if count else None)
+
+    if count:
+        assert run.spec.retry.backoff.base_delay == (
+            base_delay if base_delay is not None else default_base_delay
+        )
+    else:
+        assert run.spec.retry.backoff is None
+    with (
+        expectation,
+    ):
+        launcher._validate_retry(runtime.kind, run.spec.retry)
+
+
+def test_validate_run_retry_runtime_kind():
+    launcher = services.api.launcher.ServerSideLauncher(
+        auth_info=mlrun.common.schemas.AuthInfo()
+    )
+    runtime = mlrun.code_to_function(
+        name="test", kind="mpijob", filename=str(func_path), handler=handler
+    )
+
+    retry = {
+        "count": 3,
+    }
+    run = mlrun.run.RunObject(
+        spec=mlrun.model.RunSpec(
+            retry=retry,
+        ),
+    )
+    with (
+        pytest.raises(
+            mlrun.errors.MLRunInvalidArgumentError,
+            match=re.escape(
+                f"Retry is not supported for runtime kind mpijob, supported kinds are: "
+                f"{mlrun.runtimes.RuntimeKinds.retriable_runtimes()}"
+            ),
+        ),
+    ):
+        launcher._validate_run(runtime, run)
